requirements.txt
streamlit
streamlit-mic-recorder
langchain
langchain-openai
langchain-community
crewai
faiss-cpu
pypdf
python-dotenv
pyyaml
flask
flask-cors
numpy
soundfile
librosa
pywhispercpp
pydub
requests
